Fix explicitAction if force==true. #133
It looks like this part of the code was not used frequently.
It crashes by every iteration, because functoCache was always nullptr.

The code now throws an exception only if the force==true. Otherwise
it does not do it. I am not sure, whether it is proper behaviour,
planned by an original author of the code, but at least I did not
change the design.